refactor redis to avoid circular import
