+++ lookup.go
@@ -16,7 +16,7 @@ package zdns
 
 import (
 	"encoding/json"
-	"github.com/kwang40/zdns/iohandlers/redis"
+	"github.com/go-redis/redis"
 	"os"
 	"strconv"
 	"strings"
@@ -216,7 +216,7 @@ func DoLookups(g *GlobalLookupFactory, c *GlobalConf) error {
 
 	inHandler := GetInputHandler(c.InputHandler)
 	outHandler := GetOutputHandler(c.OutputHandler)
-	redisHandler := new(redis.RedisOutputHandler)
+	redisHandler := new(RedisOutputHandler)
 	inHandler.Initialize(c)
 	outHandler.Initialize(c)
 
@@ -279,3 +279,73 @@ func DoLookups(g *GlobalLookupFactory, c *GlobalConf) error {
 	}
 	return nil
 }
+
+
+type RedisOutputHandler struct {
+	client *redis.Client
+}
+
+func (h *RedisOutputHandler) Initialize(conf GlobalConf) {
+	h.client = redis.NewClient(&redis.Options{
+		Addr:     conf.RedisServerUrl,
+		Password: conf.RedisServerPass,
+		DB:       conf.RedisServerDB,
+	})
+}
+
+func (h *RedisOutputHandler) WriteResults(results <-chan Result, wg *sync.WaitGroup) error {
+	defer (*wg).Done()
+
+	for r := range results {
+		res, ok := r.Data.(MiekgResult)
+		if !ok {
+			o, err := json.Marshal(res)
+			if err != nil {
+				log.Fatal(err)
+			}
+			log.Warn("unable to parse result: ", string(o))
+		}
+
+		for _, a := range res.Answers {
+			var key, domain string
+			if miekgAnswer, ok := a.(MiekgAnswer); ok {
+				if miekgAnswer.Type == dns.TypeToString[dns.TypeA] || miekgAnswer.Type == dns.TypeToString[dns.TypeAAAA] {
+					key = miekgAnswer.Answer
+					domain = miekgAnswer.Name
+				}
+			} else {
+			//} else if mxAnswer, ok := a.(miekg.MXAnswer); ok {
+			//	key = mxAnswer.Answer.Answer
+			//	domain = mxAnswer.Answer.Name
+			//} else {
+				log.Warn("unimplemented answer type (not MiekgAnswer or MXAnswer")
+			}
+
+			var value []string
+			redisValue, err := h.client.Get(key).Result()
+			if err == redis.Nil { // no key found
+				value = make([]string, 0)
+			} else if err != nil {
+				log.Fatal("unable to get key:", err)
+			} else {
+				err = json.Unmarshal([]byte(redisValue), &value)
+				if err != nil {
+					log.Error("error unmarshalling redis string:", err)
+				}
+			}
+
+			value = append(value, domain)
+			jsonBytes, err := json.Marshal(value)
+			if err != nil {
+				log.Error("error marshalling redis:", err)
+			}
+			err = h.client.Set(key, string(jsonBytes[:]), 0).Err()
+			if err != nil {
+				log.Error("unable to set redis key:", err)
+			}
+		}
+
+	}
+	return nil
+}
+
